changed phone number from int to bigint. forgot about overflow...

+++ createTables.sql
@@ -4,7 +4,7 @@ CREATE TABLE person (
 	County varchar(25) NOT NULL,
 	FacebookUsername varchar(25),
 	Email varchar(25) NOT NULL,
-	PhoneNum Integer(10) NOT NULL,
+	PhoneNum bigint(10) NOT NULL,
 	MonthlyGrossIncome Integer NOT NULL,
 	NumInHousehold Integer NOT NULL,
 	ReceiveFoodStamp BIT NOT NULL,
